+++ certbot-dns-sakuracloud/certbot_dns_sakuracloud/_internal/__init__.py
@@ -1,0 +1,1 @@
+Internal implementation of `~certbot_dns_sakuracloud.dns_sakuracloud` plugin.
