+++ certbot-dns-sakuracloud/docs/api/dns_sakuracloud.rst
@@ -1,5 +1,5 @@
-:mod:`certbot_dns_sakuracloud.dns_sakuracloud`
+:mod:`certbot_dns_sakuracloud._internal.dns_sakuracloud`
 ----------------------------------------------
 
-.. automodule:: certbot_dns_sakuracloud.dns_sakuracloud
+.. automodule:: certbot_dns_sakuracloud._internal.dns_sakuracloud
    :members:
